+++ PlexCleaner/ProcessDriver.cs
@@ -386,7 +386,6 @@ public static class ProcessDriver
                 }
 
                 // Remove cover art in video tracks
-                // TODO: mediaInfoProps was commented out, is this a change in behavior?
                 _ = mediaInfoProps.Video.RemoveAll(track => track.CoverArt);
                 _ = ffProbeProps.Video.RemoveAll(track => track.CoverArt);
                 _ = mkvMergeProps.Video.RemoveAll(track => track.CoverArt);