+++ PlexCleaner/ToolsOptions.cs
@@ -24,20 +24,11 @@ public record ToolsOptions1
     public void SetDefaults()
     {
         // Set defaults based on OS
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            UseSystem = false;
-            RootPath = @".\Tools\";
-            RootRelative = true;
-            AutoUpdate = true;
-        }
-        else
-        {
-            UseSystem = true;
-            RootPath = string.Empty;
-            RootRelative = false;
-            AutoUpdate = false;
-        }
+        bool isWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
+        UseSystem = !isWindows;
+        AutoUpdate = isWindows;
+        RootPath = @".\Tools\";
+        RootRelative = true;
     }
 
     public bool VerifyValues()